+++ ios/Plugin/Plugin.m
@@ -11,6 +11,8 @@
 // Define the plugin using the CAP_PLUGIN Macro, and
 // each method the plugin supports using the CAP_PLUGIN_METHOD macro.
 CAP_PLUGIN(CapacitorDataStorageSqlite, "CapacitorDataStorageSqlite",
+           CAP_PLUGIN_METHOD(openStore, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(setTable, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(set, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(get, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(remove, CAPPluginReturnPromise);