+++ src/definitions.ts
@@ -6,20 +6,32 @@ declare module "@capacitor/core" {
 
 export interface CapacitorDataStorageSqlitePlugin {
   /**
-   * Store a data with given key and value
-   * @param {capDataStorageOptions} options 
+   * Open a store
+   * @param {capOpenStorageOptions} options {database: string, table: string}
    * @returns {Promise<capDataStorageResult>} {result:boolean}
    */
-  set(options: capDataStorageOptions): Promise<capDataStorageResult>;
+  openStore(options: capOpenStorageOptions): Promise<capDataStorageResult>;
+  /**
+   * Set or Add a table to an existing store
+   * @param {capOpenStorageOptions} options 
+   * @returns {Promise<capDataStorageResult>} {result:boolean}
+   */
+  setTable(options: capOpenStorageOptions): Promise<capDataStorageResult>;
+  /**
+   * Store a data with given key and value
+   * @param {capDataStorageOptions} options {table: string}
+   * @returns {Promise<capDataStorageResult>} {result:boolean}
+   */
+   set(options: capDataStorageOptions): Promise<capDataStorageResult>;
   /**
    * Retrieve a data value for a given data key
-   * @param {capDataStorageOptions} options 
+   * @param {capDataStorageOptions} options {key:"foo",value:"foovalue"}
    * @returns {Promise<capDataStorageResult>} {value:string}
    */
   get(options: capDataStorageOptions): Promise<capDataStorageResult>;
   /**
    * Remove a data with given key
-   * @param {capDataStorageOptions} options 
+   * @param {capDataStorageOptions} options {key:"foo"}
    * @returns {Promise<capDataStorageResult>} {result:boolean}
    */
   remove(options: capDataStorageOptions): Promise<capDataStorageResult>;
@@ -30,10 +42,15 @@ export interface CapacitorDataStorageSqlitePlugin {
   clear(): Promise<capDataStorageResult>;
   /**
    * Check if a data key exists
-   * @param {capDataStorageOptions} options 
+   * @param {capDataStorageOptions} options {key:"foo"}
    * @returns {Promise<capDataStorageResult>} {result:boolean}
    */
-  iskey(options: capDataStorageOptions): Promise<capDataStorageResult>;
+  iskey(options: capDataStorageOptions): Promise<capDataStorageResult>; 
+  /**
+   * Remove a data key if exists
+   * @param {capDataStorageOptions} options {key:"foo"}
+   * @returns {Promise<capDataStorageResult>} {result:boolean}
+   */
   remove(options: capDataStorageOptions): Promise<capDataStorageResult>;
   /**
    * Get the data key list
@@ -51,7 +68,21 @@ export interface CapacitorDataStorageSqlitePlugin {
    */
   keysvalues(): Promise<capDataStorageResult>;
 }
+export interface capOpenStorageOptions {
+  /**
+   * The storage database name
+   */
+  database?: string;  // default: 
+                      //  ios, android: storageSQLite
+                      //  web : storageIDB
+  /**
+   * The storage table name
+   */
+  table?: string;   // default:
+                    //  ios, android: storage_table
+                    //  web: storage_store
 
+}
 export interface capDataStorageOptions {
   /**
    * The data name
@@ -84,4 +115,8 @@ export interface capDataStorageResult {
    * the data keys/values list as an Array of {key:string,value:string}
    */
   keysvalues?: Array<any>;
+  /**
+   * a message
+   */
+  message?: string;
 }